Copula volume on hyper-boxes

#' u must be piecewise smaller than v, otherwise the function will return an error.

vCopula(u, v, copula, ...)

Arguments

u
numeric matrix : minimum point of the hyper-rectangles, one row per observation.

v
numeric matrix : maximum point of the hyper-rectangle, one row per observation.

copula
the copula to calcule it's measure on [u,v]

...
other parameter to be passed to methods for this generic.

Value

the measure of the copula

Details

A method is currently implemented for the main virtual class 'Copula', but it assumes that a pCopula method is avaliable for the given copula.

This function calculates the measure of the copula according to the algorythme proposed by : Umberto Cherubini & Silvia Romagnoli (2009) Computing the Volume of n-Dimensional Copulas, Applied Mathematical Finance, 16:4, 307-314, DOI: 10.1080/13504860802597311 link : http://dx.doi.org/10.1080/13504860802597311

Examples

# For a simple one-dimentional input :
cop = copula::archmCopula('Clayton',0.7,3)
vCopula(rep(0,3),rep(1,3),cop)

#> [1] 1


# the function is vectorised :
v=matrix(seq(0,1,length.out=12),ncol=3)
u=matrix(rep(0,12),ncol=3)
vCopula(u,v,cop)

#> [1] 0